Fix test dependency on bower folder! 🔨😇
With the upgrade to Polymer v2.0, I removed the paper-elements component from the dummy app, because it is no longer supported in 2.0. My tests however assumed to find paper-elements in the bower folder. Now the tests are no longer dependent on the bower folder. We use fixture bower packages now! 🍺 Cheers